finished fixing errors

+++ src/Screens/AllProperties.tsx
@@ -8,16 +8,14 @@ import {SwipeListView} from "react-native-swipe-list-view";
 import { View, StyleSheet} from "react-native";
 import BackButton from "../Components/BackButton";
 import TrashButton from "../Components/TrashButton";
+import _ from "lodash";
 
 function AllProperties() {
 	const appContext = useAppContext();
 
-	const handleDeleteProperty = useCallback(async (propertyId:number) => {
-		try{
-			await appContext.deleteProperty(propertyId);
-		}catch (e){
-			return;
-		}
+	const handleDeleteProperty = useCallback(async (propertyId?:number) => {
+		if (_.isUndefined(propertyId)) return;
+		await appContext.deleteProperty(propertyId);
 	}, [appContext]);
 
 	return(
